requirements.txt
anthropic>=0.42.0
functions-framework==3.*
google-cloud-bigquery>=3.27.0